requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
idna==3.6
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==23.2
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pytest==7.4.4
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
shellingham==1.5.4
sniffio==1.3.0
starlette==0.35.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
